# Benchmark and simple optimization for INPC accessor.