## What does the pull request do?
<!--- Give a bit of background on the PR here, together with links to with related issues etc. -->
Fixes low hanging optimization fruit in INPC accessor :)

## What is the current behavior?
<!--- If the PR is a fix, describe the current incorrect behavior, otherwise delete this section. -->

|       Method |     Mean |   Error |  StdDev |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|------------- |---------:|--------:|--------:|-------:|------:|------:|----------:|
| InpcAccessor | 117.8 ns | 2.08 ns | 3.42 ns | 0.0477 |     - |     - |     200 B |
## What is the updated/expected behavior with this PR?
<!--- Describe how to test the PR. -->
|       Method |     Mean |    Error |   StdDev |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|------------- |---------:|---------:|---------:|-------:|------:|------:|----------:|
| InpcAccessor | 62.66 ns | 0.525 ns | 0.491 ns | 0.0114 |     - |     - |      48 B |

## How was the solution implemented (if it's not obvious)?
<!--- Include any information that might be of use to a reviewer here. -->
Just use optimized runtime function to query for properties. Own caching is not really worth it here (was like 10 ns faster in best case scenario).

## Checklist

- [ ] Added unit tests (if possible)?
- [ ] Added XML documentation to any related classes?
- [ ] Consider submitting a PR to https://github.com/AvaloniaUI/Avaloniaui.net with user documentation